(PUP-7559) Use the file ensure property to get default selinux context
Previously, function get_selinux_default_context in Puppet::Util::SELinux
would try to fetch the mode using lstat, to pass it over to matchpathcon.
This only worked when the file exists, and in any other case we pass 0 as
value, which may result in an incorrect context when a directory or
symlink is desired, breaking idempotency.

This patch uses the ensure property of the File resource to do a better
guess, only if the file does not exist already.

NOTE: unit tests pending